Remove all custom flake8 configuration
This is a new code base, so there's no need to ignore things. I was only
going to sync with Matplotlib's config, but none of the existing code
violated any of the additional ignores, and we only have those in
Matplotlib to prevent mass changes.

Also, we should definitely not ignore `__init__.py`, as that's the only
file with code in it.