Make the documentation build reproducibly
Whilst working on the Reproducible Builds effort [0], I noticed that
mpl-sphinx-theme could not be built reproducibly.

This is because the documentation embedded the current build year. This patch
changes the behaviour of the Sphinx documentation so that it can that
optionally use SOURCE_DATE_EPOCH [1] as the source of this value instead.

I originally filed this in Debian as bug #1005826 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/1005826